Update handler, use new on message method, add more commands
The handler automatically picks up all commands from the ./commands directory and inserts them into an object. They are then accessed in the on normal chat message type event to check if a command should be triggered.

The types are available in client.types as an object containing all chat and role types.